requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.4.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
pyyaml>=6.0
python-dotenv>=1.0.0
prometheus-client>=0.18.0
structlog>=23.2.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
websockets>=12.0
httpx>=0.25.0
cachetools>=5.3.0
python-multipart>=0.0.6
email-validator>=2.1.0
psutil>=5.9.0
Web3>=7.13.0
ccxt>=4.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
flake8
# Testing dependencies
pytest>=8.4.0
pytest-asyncio>=1.2.0
# Frontend dependencies (installed via npm)
# lucide-react (installed via npm install lucide-react)
# Additional dependencies installed during Agent B setup:
# redis-server (via apt install)
# pytest (already in requirements.txt, installed via pip3 install)
# All Python packages from requirements.txt successfully installed

# Additional dependencies installed during implementation:
# flake8>=7.3.0 (for linting)
# redis>=6.4.0 (for Redis integration)
pytz>=2023.3
# All core dependencies successfully installed and tested

# Agent A Implementation Dependencies Added:
# Frontend testing dependencies (installed via npm):
# - @testing-library/react@^16.3.0
# - @testing-library/jest-dom@^6.9.1  
# - @testing-library/user-event@^14.6.1
# - vitest@^3.2.4
# - jsdom@^27.0.0
# 
# Backend API enhancements:
# - Added new mode config endpoints in strategies.py
# - Enhanced error code system across all Agent A components
# - Updated EventDrivenStrategyEngine method signatures

# Additional dependencies found in actual codebase imports:
# Core Python libraries (already included in standard library):
# - asyncio, json, logging, pathlib, datetime, typing, dataclasses, enum, decimal, uuid, os, sys

# Third-party dependencies actually used in code:
requests>=2.31.0
aiofiles>=23.2.0
python-dateutil>=2.8.2
cryptography>=41.0.0
eth-account>=0.9.0
eth-utils>=2.3.0
hexbytes>=0.3.1
typing-extensions>=4.8.0

# Dependencies already present but confirmed in use:
# - fastapi (API routes)
# - uvicorn (server)
# - pandas (data processing)
# - numpy (numerical operations)
# - redis (caching/storage)
# - pydantic (data validation)
# - structlog (structured logging)
# - yaml (configuration)
# - pytz (timezone handling)
# - Web3 (blockchain interactions)
# - ccxt (crypto exchange APIs)

# Frontend dependencies (package.json EXISTS and is properly configured):
# ✅ CONFIRMED: frontend/package.json contains all necessary dependencies
# 
# 🚨 IMPORTANT: Frontend dependencies are managed by NPM, NOT pip!
# To install frontend dependencies, use:
#   cd frontend && npm install
# 
# Key dependencies from actual package.json:
# - react (^19.1.1) - React framework
# - react-dom (^19.1.1) - React DOM rendering
# - lucide-react (^0.544.0) - Icon library
# - plotly.js-dist-min (^3.1.1) - Chart library
# - vite (^5.3.1) - Build tool
# - typescript (^5.2.2) - Type checking
# - @vitejs/plugin-react (^4.3.1) - Vite React plugin
# - vitest (^3.2.4) - Testing framework
# - @testing-library/react (^16.3.0) - React testing utilities
# - @testing-library/jest-dom (^6.9.1) - Jest DOM matchers
# - @testing-library/user-event (^14.6.1) - User interaction testing
# - jsdom (^27.0.0) - DOM environment for testing
# - Additional UI libraries: @radix-ui/*, @tanstack/react-query, recharts, etc.
